requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
debugpy==1.6.7.post1
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
fastcore==1.5.29
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyterlab==4.1.0a1
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
lightning-utilities==0.9.0
llvmlite==0.41.1
Markdown==3.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prettytable==3.8.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.8.4
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scikit-learn-extra==0.3.0
scipy==1.11.2
seaborn==0.12.2
sentry-sdk==1.37.1
setproctitle==1.3.3
shap==0.44.0
six==1.16.0
sklearn==0.0.post7
slicer==0.0.7
smmap==5.0.1
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.0.1
torch-explain==1.5.1
torchaudio==2.0.2
torchmetrics==1.1.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.16
wandb==0.16.0
wcwidth==0.2.6
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
